internal: rename TrackNames to Names
